requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
flake8==7.1.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
pillow==11.1.0
psycopg==3.2.6
pycodestyle==2.12.1
pyflakes==3.2.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
uwsgi==2.0.29
python-decouple==3.8